# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Police Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /soldiers:
    get:
      operationId: listSoldiers
      tags: [soldiers]
      summary: List all soldiers with status and stats
      responses:
        "200":
          description: List of soldiers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SoldierWithStatus"
    post:
      operationId: createSoldier
      tags: [soldiers]
      summary: Add a new soldier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSoldierBody"
      responses:
        "201":
          description: Soldier created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Soldier"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /soldiers/{id}:
    get:
      operationId: getSoldier
      tags: [soldiers]
      summary: Get a soldier by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Soldier details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SoldierWithStatus"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateSoldier
      tags: [soldiers]
      summary: Update a soldier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSoldierBody"
      responses:
        "200":
          description: Updated soldier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Soldier"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteSoldier
      tags: [soldiers]
      summary: Delete a soldier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /warnings:
    get:
      operationId: listWarnings
      tags: [warnings]
      summary: List all warnings (admin only)
      responses:
        "200":
          description: List of warnings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Warning"
    post:
      operationId: createWarning
      tags: [warnings]
      summary: Issue a warning to a soldier (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWarningBody"
      responses:
        "201":
          description: Warning created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Warning"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /warnings/{id}:
    delete:
      operationId: deleteWarning
      tags: [warnings]
      summary: Delete a warning (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /warnings/soldier/{soldierId}:
    get:
      operationId: getSoldierWarnings
      tags: [warnings]
      summary: Get warnings for a specific soldier
      parameters:
        - name: soldierId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Soldier warnings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Warning"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Soldier:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        age:
          type: integer
        unit:
          type: string
        rank:
          type: string
        username:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - age
        - unit
        - rank
        - createdAt
    SoldierWithStatus:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        age:
          type: integer
        unit:
          type: string
        rank:
          type: string
        username:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        isCheckedIn:
          type: boolean
        totalMinutes:
          type: integer
        warningsCount:
          type: integer
        lastCheckIn:
          type: ["string", "null"]
          format: date-time
        lastCheckOut:
          type: ["string", "null"]
          format: date-time
      required:
        - id
        - name
        - age
        - unit
        - rank
        - createdAt
        - isCheckedIn
        - totalMinutes
        - warningsCount
    CreateSoldierBody:
      type: object
      properties:
        name:
          type: string
        age:
          type: integer
        unit:
          type: string
        rank:
          type: string
        username:
          type: string
        password:
          type: string
      required:
        - name
        - age
        - unit
        - rank
    UpdateSoldierBody:
      type: object
      properties:
        name:
          type: string
        age:
          type: integer
        unit:
          type: string
        rank:
          type: string
        username:
          type: string
        password:
          type: string
...
    Warning:
      type: object
      properties:
        id:
          type: integer
        soldierId:
          type: integer
        soldierName:
          type: string
        reason:
          type: string
        givenBy:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - soldierId
        - soldierName
        - reason
        - givenBy
        - createdAt
    CreateWarningBody:
      type: object
      properties:
        soldierId:
          type: integer
        reason:
          type: string
      required:
        - soldierId
        - reason
...